# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Consumer:
    properties:
      ack_policy:
        $ref: '#/definitions/models.ConsumerAckPolicy'
      ack_wait:
        type: integer
      backoff:
        items:
          type: integer
        type: array
      consumer_type:
        $ref: '#/definitions/models.ConsumerType'
      created_at:
        $ref: '#/definitions/models.CustomTime'
      deliver_group:
        type: string
      deliver_policy:
        $ref: '#/definitions/models.ConsumerDeliverPolicy'
      deliver_subject:
        type: string
      description:
        type: string
      durable:
        type: string
      filter_subject:
        type: string
      filter_subjects:
        items:
          type: string
        type: array
      flow_control:
        type: boolean
      headers_only:
        type: boolean
      id:
        type: string
      idle_heartbeat:
        type: integer
      inactive_threshold:
        type: integer
      jetstream:
        $ref: '#/definitions/models.JetStream'
      jetstream_id:
        type: string
      max_ack_pending:
        type: integer
      max_batch:
        type: integer
      max_bytes:
        type: integer
      max_deliver:
        type: integer
      max_expires:
        type: integer
      max_request_batch:
        type: integer
      max_request_expires:
        type: integer
      max_request_max_bytes:
        type: integer
      max_waiting:
        type: integer
      memory_storage:
        type: boolean
      metadata:
        $ref: '#/definitions/models.MetadataMap'
      mode:
        type: string
      name:
        type: string
      nats_operate_user:
        $ref: '#/definitions/models.User'
      nats_operate_user_id:
        type: string
      opt_start_seq:
        type: integer
      opt_start_time:
        type: string
      pause_until:
        type: string
      pinned_ttl:
        type: integer
      priority_groups:
        items:
          type: string
        type: array
      priority_policy:
        $ref: '#/definitions/models.ConsumerPriorityPolicy'
      rate_limit:
        type: integer
      rate_limit_bps:
        type: integer
      replay_policy:
        $ref: '#/definitions/models.ConsumerReplayPolicy'
      replicas:
        type: integer
      retry_policy:
        $ref: '#/definitions/models.ConsumerRetryPolicy'
      sample_frequency:
        type: string
      start_seq:
        type: integer
      start_time:
        type: string
      stats:
        $ref: '#/definitions/models.ConsumerStats'
      status:
        type: string
      sync_message:
        type: string
      sync_status:
        $ref: '#/definitions/models.ConsumerSyncStatus'
      updated_at:
        $ref: '#/definitions/models.CustomTime'
    type: object
  models.ConsumerAckPolicy:
    enum:
    - AckExplicit
    - AckAll
    - AckNone
    type: string
    x-enum-varnames:
    - AckExplicit
    - AckAll
    - AckNone
  models.ConsumerConfigDiff:
    properties:
      cluster_config:
        $ref: '#/definitions/models.ConsumerConfigView'
      database_config:
        $ref: '#/definitions/models.ConsumerConfigView'
      differences:
        items:
          $ref: '#/definitions/models.ConfigDifference'
        type: array
      has_difference:
        type: boolean
    type: object
...
  models.ConsumerDeliverPolicy:
    enum:
    - all
    - last
    - new
    - by_start_sequence
    - by_start_time
    - last_per_subject
    type: string
    x-enum-varnames:
    - DeliverAll
    - DeliverLast
    - DeliverNew
    - DeliverByStartSeq
    - DeliverByStartTime
    - DeliverLastPerSubject
  models.ConsumerPriorityPolicy:
    enum:
    - ""
    - pinned
    type: string
    x-enum-varnames:
    - PriorityNone
    - PriorityPinned
  models.ConsumerReplayPolicy:
    enum:
    - instant
    - original
    type: string
    x-enum-varnames:
    - ReplayInstant
    - ReplayOriginal
  models.ConsumerRetryPolicy:
    enum:
    - exponential
    - uniform
    type: string
    x-enum-varnames:
    - RetryExponential
    - RetryUniform
  models.ConsumerStats:
    properties:
      ack_floor_seq:
        type: integer
      ack_floor_stream:
        type: integer
      cluster:
        $ref: '#/definitions/models.ClusterInfo'
      consumer_name:
        type: string
      created:
        $ref: '#/definitions/models.CustomTime'
      delivered_seq:
        type: integer
      delivered_stream:
        type: integer
      delivered_time:
        $ref: '#/definitions/models.CustomTime'
      num_ack_pending:
        type: integer
      num_pending:
        type: integer
      num_redelivered:
        type: integer
      num_waiting:
        type: integer
      pause_remaining:
        example: 0
        type: integer
      paused:
        type: boolean
      push_bound:
        type: boolean
      stream_name:
        type: string
      timestamp:
        $ref: '#/definitions/models.CustomTime'
    type: object
  models.ConsumerSyncStatus:
    enum:
    - pending
    - syncing
    - synced
    - failed
    type: string
    x-enum-varnames:
    - ConsumerSyncPending
    - ConsumerSyncSyncing
    - ConsumerSyncSynced
    - ConsumerSyncFailed
  models.ConsumerType:
    enum:
    - pull
    - push
    type: string
    x-enum-varnames:
    - ConsumerTypePull
    - ConsumerTypePush
...
  models.CreateConsumerRequest:
    properties:
      ack_policy:
        $ref: '#/definitions/models.ConsumerAckPolicy'
      ack_wait:
        type: string
      backoff:
        items:
          type: string
        type: array
      consumer_name:
        type: string
      consumer_type:
        $ref: '#/definitions/models.ConsumerType'
      deliver_group:
        type: string
      deliver_policy:
        $ref: '#/definitions/models.ConsumerDeliverPolicy'
      deliver_subject:
        type: string
      description:
        type: string
      durable:
        type: string
      filter_subject:
        type: string
      filter_subjects:
        items:
          type: string
        type: array
      flow_control:
        type: boolean
      headers_only:
        type: boolean
      idle_heartbeat:
        type: string
      inactive_threshold:
        type: integer
      jetstream_id:
        type: string
      max_ack_pending:
        type: integer
      max_batch:
        type: integer
      max_bytes:
        type: integer
      max_deliver:
        type: integer
      max_expires:
        type: string
      max_request_batch:
        type: integer
      max_request_expires:
        type: integer
      max_request_max_bytes:
        type: integer
      max_waiting:
        type: integer
      memory_storage:
        type: boolean
      metadata:
        additionalProperties:
          type: string
        type: object
      mode:
        enum:
        - push
        - pull
        type: string
      name:
        type: string
      nats_operate_user_id:
        type: string
      opt_start_seq:
        type: integer
      opt_start_time:
        type: string
      rate_limit:
        type: integer
      rate_limit_bps:
        type: integer
      replay_policy:
        $ref: '#/definitions/models.ConsumerReplayPolicy'
      replicas:
        type: integer
      retry_policy:
        $ref: '#/definitions/models.ConsumerRetryPolicy'
      sample_freq:
        type: string
      start_seq:
        type: integer
      start_time:
        type: string
      stream_name:
        type: string
    required:
    - ack_policy
    - consumer_name
    - mode
    - stream_name
    type: object
...
  models.CustomTime:
    properties:
      time.Time:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.JSONMap:
    additionalProperties: true
    type: object
...
  models.JetStream:
    properties:
      allow_direct:
        type: boolean
      allow_rollup_hdrs:
        type: boolean
      bytes_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 当前字节数单位
      bytes_value:
        description: 当前字节数值
        type: number
      cluster:
        $ref: '#/definitions/models.Cluster'
      cluster_id:
        type: string
      compression:
        $ref: '#/definitions/models.JetStreamCompressionType'
      created_at:
        allOf:
        - $ref: '#/definitions/models.CustomTime'
        description: 时间戳
      deny_delete:
        type: boolean
      deny_purge:
        type: boolean
      description:
        type: string
      discard:
        $ref: '#/definitions/models.JetStreamDiscardPolicy'
      discard_new_per_subject:
        description: 高级配置
        type: boolean
      duplicate_window:
        description: 重复检测窗口（秒）
        type: integer
      first_seq:
        description: 第一个序列号
        type: integer
      id:
        type: string
      last_seq:
        description: 最后一个序列号
        type: integer
      max_age:
        description: 消息最大存活时间（秒）
        type: integer
      max_bytes_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 最大字节数单位
      max_bytes_value:
        description: 最大字节数值
        type: number
      max_consumers:
        description: 最大消费者数量
        type: integer
      max_msg_size_unit:
        allOf:
        - $ref: '#/definitions/models.StorageUnit'
        description: 最大单条消息大小单位
      max_msg_size_value:
        description: 最大单条消息大小数值
        type: number
      max_msgs:
        description: 限制配置
        type: integer
      max_msgs_per_subject:
        description: 每个主题的最大消息数量
        type: integer
      messages:
        description: 统计信息
        type: integer
      metadata:
        $ref: '#/definitions/models.MetadataMap'
      mirror_direct:
        type: boolean
      name:
        type: string
      nats_operate_user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联的NATS操作用户信息（不存储在数据库中）
      nats_operate_user_id:
        description: 重构字段：操作NATS的用户身份
        type: string
      no_ack:
        description: 是否禁用确认
        type: boolean
      num_consumers:
        description: 当前消费者数量
        type: integer
      placement_cluster:
        description: 放置标签和元数据
        type: string
      placement_tags:
        items:
          type: string
        type: array
      replicas:
        description: 副本和持久化配置
        type: integer
      retention:
        $ref: '#/definitions/models.JetStreamRetentionPolicy'
      sealed:
        description: 是否密封流
        type: boolean
      status:
        description: active, inactive, error
        type: string
      storage:
        $ref: '#/definitions/models.JetStreamStorageType'
      subjects:
        description: 流配置
        items:
          type: string
        type: array
      sync_message:
        type: string
      sync_status:
        allOf:
        - $ref: '#/definitions/models.JetStreamSyncStatus'
        description: 新增：同步状态管理
      updated_at:
        $ref: '#/definitions/models.CustomTime'
    type: object
...
  models.MetadataMap:
    additionalProperties:
      type: string
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.UpdateConsumerRequest:
    properties:
      ack_policy:
        $ref: '#/definitions/models.ConsumerAckPolicy'
      ack_wait:
        type: string
      backoff:
        items:
          type: string
        type: array
      consumer_name:
        type: string
      consumer_type:
        $ref: '#/definitions/models.ConsumerType'
      deliver_group:
        type: string
      deliver_policy:
        $ref: '#/definitions/models.ConsumerDeliverPolicy'
      deliver_subject:
        type: string
      description:
        type: string
      durable:
        type: string
      filter_subject:
        type: string
      filter_subjects:
        items:
          type: string
        type: array
      flow_control:
        type: boolean
      headers_only:
        type: boolean
      idle_heartbeat:
        type: string
      inactive_threshold:
        type: integer
      jetstream_id:
        type: string
      max_ack_pending:
        type: integer
      max_batch:
        type: integer
      max_bytes:
        type: integer
      max_deliver:
        type: integer
      max_expires:
        type: string
      max_waiting:
        type: integer
      metadata:
        additionalProperties:
          type: string
        type: object
      mode:
        type: string
      name:
        type: string
      rate_limit:
        type: integer
      rate_limit_bps:
        type: integer
      replay_policy:
        $ref: '#/definitions/models.ConsumerReplayPolicy'
      replicas:
        type: integer
      retry_policy:
        $ref: '#/definitions/models.ConsumerRetryPolicy'
      sample_freq:
        type: string
      stream_name:
        type: string
    type: object
...
  models.User:
    properties:
      account_id:
        type: string
      created_at:
        $ref: '#/definitions/models.CustomTime'
      creds_file:
        type: string
      description:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      jwt_claims:
        $ref: '#/definitions/models.JSONMap'
      jwt_text:
        description: 存储实际的JWT密文
        type: string
      limits:
        $ref: '#/definitions/models.UserLimits'
      n_key:
        type: string
      name:
        type: string
      permissions:
        $ref: '#/definitions/models.UserPermissions'
      public_key:
        type: string
      status:
        $ref: '#/definitions/models.UserStatus'
      updated_at:
        $ref: '#/definitions/models.CustomTime'
    required:
    - account_id
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing NATS control plane with multi-cluster support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NATS Control API
  version: "1.0"
paths:
...
  /consumers:
    get:
      consumes:
      - application/json
      description: Retrieve a list of Consumers with optional filtering and pagination
      parameters:
      - description: Filter by Cluster ID
        in: query
        name: cluster_id
        type: string
      - description: Filter by JetStream ID
        in: query
        name: jetstream_id
        type: string
      - description: Filter by NATS operate user ID
        in: query
        name: nats_operate_user_id
        type: string
      - description: Filter by consumer type (pull, push)
        in: query
        name: consumer_type
        type: string
      - description: Filter by status (active, paused)
        in: query
        name: status
        type: string
      - description: Filter by sync status (pending, synced, failed)
        in: query
        name: sync_status
        type: string
      - description: Search by name or description
        in: query
        name: search
        type: string
      - description: Filter by durability (true, false)
        in: query
        name: is_durable
        type: boolean
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Consumer'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List Consumers
      tags:
      - Consumer
    post:
      consumes:
      - application/json
      description: Create a new Consumer on the specified JetStream
      parameters:
      - description: Consumer configuration
        in: body
        name: consumer
        required: true
        schema:
          $ref: '#/definitions/models.CreateConsumerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Consumer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new Consumer
      tags:
      - Consumer
  /consumers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Consumer from the system and NATS cluster
      parameters:
      - description: Consumer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Consumer deleted successfully
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a Consumer
      tags:
      - Consumer
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific Consumer
      parameters:
      - description: Consumer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Consumer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get a Consumer
      tags:
      - Consumer
    put:
      consumes:
      - application/json
      description: Update configuration of an existing Consumer
      parameters:
      - description: Consumer ID
        in: path
        name: id
        required: true
        type: string
      - description: Consumer update configuration
        in: body
        name: consumer
        required: true
        schema:
          $ref: '#/definitions/models.UpdateConsumerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Consumer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a Consumer
      tags:
      - Consumer
  /consumers/{id}/diff:
    get:
      consumes:
      - application/json
      description: Compare Consumer configuration between database and NATS cluster
        to detect differences
      parameters:
      - description: Consumer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConsumerConfigDiff'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Compare Consumer configuration
      tags:
      - Consumer
  /consumers/{id}/pause:
    post:
      consumes:
      - application/json
      description: Pause message delivery for a Consumer
      parameters:
      - description: Consumer ID
        in: path
        name: id
        required: true
        type: string
      - description: Pause until timestamp (optional)
        in: body
        name: pause_until
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Pause a Consumer
      tags:
      - Consumer
  /consumers/{id}/resume:
    post:
      consumes:
      - application/json
      description: Resume message delivery for a paused Consumer
      parameters:
      - description: Consumer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Resume a Consumer
      tags:
      - Consumer
  /consumers/{id}/retry:
    post:
      consumes:
      - application/json
      description: Retry synchronization for a Consumer that failed to sync to NATS
      parameters:
      - description: Consumer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Retry Consumer sync
      tags:
      - Consumer
  /consumers/stats:
    get:
      consumes:
      - application/json
      description: Get aggregated statistics for all Consumers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get Consumer statistics
      tags:
      - Consumer
...
schemes:
- http
- https
swagger: "2.0"